packages/pyquipu-common/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pyquipu-common"
version = "0.1.1"
authors = [{ name="doucx", email="doucxldh@gmail.com" }]
description = "Common utilities and helpers for the Quipu ecosystem."
requires-python = ">=3.10"
license = "Apache-2.0"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
keywords = ["quipu", "devops", "common", "utils"]

dependencies = [
    "pyneedle-bus ~= 0.1.4",
]

[project.urls]
Homepage = "https://github.com/quipu-dev/Quipu"
Repository = "https://github.com/quipu-dev/Quipu"
"Bug Tracker" = "https://github.com/quipu-dev/Quipu/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/quipu"]